Made sure that
>>> from pysqlite2 import test
>>> from pysqlite2 import dbapi2 as sqlite

etc. don't work from the root directory of the source tree.

I believe we had at least 10 instances of people trying to run the tests from
the source tree, which always failed because the compiled "_sqlite.so"
extension could not be found there. Now this will work once pysqlite is
installed - and people will not stumble over problem this any longer.